extract v2 controller to separated implementation. Fix issues with validation errors when receiving correct request
