+++ local-rest-scorer/src/main/java/ai/h2o/mojos/deploy/local/rest/config/WebConfig.java
@@ -1,0 +1,15 @@
+package ai.h2o.mojos.deploy.local.rest.config;
+
+import ai.h2o.mojos.deploy.local.rest.converters.MultipartConverter;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+  @Override
+  public void addFormatters(FormatterRegistry registry) {
+    registry.addConverter(new MultipartConverter());
+  }
+}
