+++ local-rest-scorer/src/test/java/ai/h2o/mojos/deploy/local/rest/controller/ModelsApiControllerTest.java
@@ -5,7 +5,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import ai.h2o.mojos.deploy.common.rest.model.CapabilityType;
-import ai.h2o.mojos.deploy.common.rest.v2.model.ScoreMediaRequest;
 import ai.h2o.mojos.deploy.common.transform.MojoScorer;
 import ai.h2o.mojos.deploy.common.transform.SampleRequestBuilder;
 import ai.h2o.mojos.deploy.common.transform.ShapleyLoadOption;
@@ -14,7 +13,6 @@ import ai.h2o.mojos.runtime.api.MojoPipelineService;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -26,8 +24,6 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.core.io.Resource;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 @ExtendWith(MockitoExtension.class)
@@ -118,22 +114,5 @@ class ModelsApiControllerTest {
 
     // Then
     assertEquals(expectedCapabilities, response.getBody());
-  }
-
-  @Test
-  void verifyScoreMedia_ReturnsUnimplemented() {
-    // Given
-    MojoScorer scorer = mock(MojoScorer.class);
-    when(scorer.getEnabledShapleyTypes()).thenReturn(ShapleyLoadOption.ALL);
-    ScoreMediaRequest request = mock(ScoreMediaRequest.class);
-    List<Resource> files = new ArrayList<>();
-    ModelsApiController controller = new ModelsApiController(scorer, sampleRequestBuilder);
-
-    // When
-    ResponseEntity<ai.h2o.mojos.deploy.common.rest.v2.model.ScoreResponse> response =
-        controller.getScoreMedia(request, files);
-
-    // Then
-    assertEquals(response.getStatusCode(), HttpStatus.NOT_IMPLEMENTED);
   }
 }
